RTL Support - swap elements
Phone numbers (like any number) should be written LTR (Left-to-Right) regardless of the language.

In the component of this library, in RTL (Right-to-Left) mode, the country code (and flag) are displayed on the right side and the phone number on the left side of the screen. 

Based on the value of an RTL property, the order of the components is changed.